requirements.txt
absl-py==0.9.0
altgraph==0.17
astor==0.8.1
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
dill==0.3.1.1
future==0.18.2
gast==0.2.2
google-auth==1.12.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
h5py==2.10.0
idna==2.9
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.2.1
nltk==3.4.5
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.0
promise==2.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyInstaller==3.6
pyparsing==2.4.6
python-dateutil==2.8.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
tensorboard==2.0.2
tensorflow-datasets==2.1.0
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
tensorflow-metadata==0.21.1
termcolor==1.1.0
tqdm==4.44.1
urllib3==1.25.8
Werkzeug==1.0.1
wrapt==1.12.1
